services/html-parser/requirements.txt
flask==3.0.0
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
markdownify==0.11.6
html5lib==1.1
